Improve naming and formatting
